requirements.txt
pandas
openpyxl
xlsxwriter
PySide6
matplotlib
fpdf2
pyinstaller